python/requirements.txt
annotated-types==0.5.0
anyio==4.0.0
bardapi==0.1.33
beautifulsoup4==4.12.2
browser-cookie3==0.19.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
deep-translator==1.11.4
exceptiongroup==1.1.3
fastapi==0.103.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
jeepney==0.8.0
langdetect==1.0.9
lz4==4.3.2
proto-plus==1.22.3
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
